app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlalchemy>=2.0.20
pyjwt>=2.8.0
python-multipart>=0.0.6
httpx>=0.25.0
websockets>=12.0.0
redis>=5.0.0
aio-pika>=9.3.0
git+https://github.com/Lawly-code/protos.git@0.1.4#egg=protos
git+https://github.com/Lawly-code/database.git@0.2.14#egg=lawly-db